+++ src/PowerModels.jl
@@ -21,6 +21,7 @@ include("core/objective.jl")
 include("core/solution.jl")
 
 include("form/acp.jl")
+include("form/act.jl")
 include("form/dcp.jl")
 include("form/wr.jl")
 include("form/wrm.jl")