+++ test/opf.jl
@@ -30,6 +30,36 @@
 end
 
 
+@testset "test ac tan opf" begin
+    @testset "3-bus case" begin
+        result = run_opf("../test/data/case3.m", ACTPowerModel, ipopt_solver)
+
+        @test result["status"] == :LocalOptimal
+        @test isapprox(result["objective"], 5812; atol = 1e0)
+    end
+    @testset "5-bus asymmetric case" begin
+        result = run_opf("../test/data/case5_asym.m", ACTPowerModel, ipopt_solver)
+
+        @test result["status"] == :LocalOptimal
+        @test isapprox(result["objective"], 17551; atol = 1e0)
+    end
+    @testset "6-bus case" begin
+        result = run_opf("../test/data/case6.m", ACTPowerModel, ipopt_solver)
+
+        @test result["status"] == :LocalOptimal
+        @test isapprox(result["objective"], 11567; atol = 1e0)
+        @test isapprox(result["solution"]["bus"]["1"]["va"], 0.0; atol = 1e-4)
+        @test isapprox(result["solution"]["bus"]["4"]["va"], 0.0; atol = 1e-4) 
+    end
+    @testset "24-bus rts case" begin
+        result = run_opf("../test/data/case24.m", ACTPowerModel, ipopt_solver)
+
+        @test result["status"] == :LocalOptimal
+        @test isapprox(result["objective"], 79804; atol = 1e0)
+    end
+end
+
+
 @testset "test dc opf" begin
     @testset "3-bus case" begin
         result = run_dc_opf("../test/data/case3.m", ipopt_solver)